requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
clip==0.2.0
dotenv==0.9.9
filelock==3.18.0
fsspec==2025.3.2
geographiclib==2.0
geojson==3.2.0
geopy==2.4.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
pillow==11.2.1
python-dotenv==1.1.0
requests==2.32.3
shapely==2.1.0
sympy==1.14.0
torch==2.7.0
typing_extensions==4.13.2
urllib3==2.4.0
